Add duke.task package
