+++ src/main/java/duke/task/Event.java
@@ -1,0 +1,34 @@
+package duke.task;
+
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.TemporalAccessor;
+
+public class Event extends Task {
+    private static String SYMBOL = "E";
+    protected TemporalAccessor startDatetime;
+    protected TemporalAccessor endDatetime;
+
+    public Event(String description, TemporalAccessor startDatetime, TemporalAccessor endDatetime) {
+        super(description);
+        this.startDatetime = startDatetime;
+        this.endDatetime = endDatetime;
+    }
+
+    @Override
+    public String getDataString() {
+        DateTimeFormatter parser = DateTimeFormatter.ofPattern("dd-MM-yyyy[ HHmm]");
+        return String.join(" | ",
+                           Event.SYMBOL, super.isDone ? "1" : "0",
+                           super.getDescription(),
+                           parser.format(startDatetime),
+                           parser.format(endDatetime));
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[%s]%s (from: %s to: %s)",
+                             Event.SYMBOL, super.toString(),
+                             super.formatDateTime(this.startDatetime),
+                             super.formatDateTime(this.endDatetime));
+    }
+}
